+++ mllam_data_prep/create_dataset.py
@@ -241,7 +241,7 @@ def create_dataset_zarr(fp_config, fp_zarr: str = None):
     """
     config = Config.from_yaml_file(file=fp_config)
 
-    expected_schema_version = "v0.2.0+dev"
+    expected_schema_version = "v0.5.0"
     assert (
         config.schema_version == expected_schema_version
     ), f"Expected schema version {expected_schema_version}, got {config.schema_version}"